app/requirements.txt
Flask==2.1.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
click==8.1.2
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
Werkzeug==2.1.1
zipp==3.8.0